+++ sdk/communication/communication-identity/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History
 
+## 1.2.0
+
+### Features Added
+- Added support to customize the Communication Identity access token's validity period:
+    - Added `getToken(user: CommunicationUserIdentifier, scopes: TokenScope[], options?: OperationOptions)` method that provide the ability to create a Communication Identity access token with custom expiration.
+    - Added `GetTokenOptions` to pass mandatory and configurable parameters when using `getToken` APIs.
+    - Added a new API version `2022-10-01` that is now the default API version.
+
 ## 1.1.0 (Unreleased)
 
 ### Features Added